n-ipv4ll: create dynamic IPv4 link-local address configuration library
This is a simple wrapper around n-acd, taking care of automatic address
selection.

Rather than accepting an IP address as input to check for, an enumeration
of the link-local subnet is given as a 64-bit number. The enumeration
should be randomly distributed, and constant for a given link.

Unreserved IP addresses in the link-local subnet is attempted until one
is found that is not already in use.

If a conflict is detected after an IP address has been announced, we
attempt to defend it once, before giving up.

Signed-off-by: Tom Gundersen <teg@jklm.no>